feat(tvix/castore): bump [Directory,File]Node size to u64
Having more than 4GiB files is quite possible (think about the NixOS
graphical installer, and an uncompressed iso of it).

No wire format changes.

Reviewed-on: https://cl.tvl.fyi/c/depot/+/9950
Tested-by: BuildkiteCI

+++ castore-go/castore.go
@@ -12,9 +12,9 @@ import (
 // The size of a directory is calculated by summing up the numbers of
 // `directories`, `files` and `symlinks`, and for each directory, its size
 // field.
-func (d *Directory) Size() uint32 {
-	var size uint32
-	size = uint32(len(d.Files) + len(d.Symlinks))
+func (d *Directory) Size() uint64 {
+	var size uint64
+	size = uint64(len(d.Files) + len(d.Symlinks))
 	for _, d := range d.Directories {
 		size += 1 + d.Size
 	}
